Allow '/' in test name (rust)
Use criterion rs
