+++ .github/workflows/criterion-rs.yml
@@ -1,4 +1,4 @@
-name: Criterion Example
+name: Criterion.rs Example
 on:
     push:
         branches:
@@ -6,19 +6,19 @@ on:
 
 jobs:
     benchmark:
-        name: Run Criterion benchmark example
+        name: Run Criterion.rs benchmark example
         runs-on: ubuntu-latest
         steps:
             - uses: actions/checkout@v2
             - run: rustup toolchain update
             - name: Run benchmark
-              run: cd examples/criterion && cargo bench | tee output.txt
+              run: cd examples/criterion-rs && cargo bench -- --output-format bencher | tee output.txt
             - name: Store benchmark result
               uses: rhysd/github-action-benchmark@v1
               with:
                   name: Rust Benchmark
-                  tool: 'criterion'
-                  output-file-path: examples/criterion/output.txt
+                  tool: 'cargo'
+                  output-file-path: examples/criterion-rs/output.txt
                   # Use personal access token instead of GITHUB_TOKEN due to https://github.community/t5/GitHub-Actions/Github-action-not-triggering-gh-pages-upon-push/td-p/26869/highlight/false
                   github-token: ${{ secrets.PERSONAL_GITHUB_TOKEN }}
                   auto-push: true
