+++ .gitignore
@@ -3,8 +3,8 @@
 /src/*.js.map
 /examples/rust/Cargo.lock
 /examples/rust/target
-/examples/criterion/Cargo.lock
-/examples/criterion/target
+/examples/criterion-rs/Cargo.lock
+/examples/criterion-rs/target
 /test/*.js
 /test/*.js.map
 /scripts/*.js
