+++ examples/criterion-rs/Cargo.toml
@@ -7,7 +7,10 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-criterion = "0.3.0"
+criterion = "*"
+
+[lib]
+bench = false
 
 [[bench]]
 name = "bench"
