+++ examples/criterion-rs/README.md
@@ -1,7 +1,6 @@
 # Rust Criterion example for benchmarking with `cargo bench`
 
--   [Workflow for this example](../../.github/workflows/rust.yml)
--   [Action log of this example](https://github.com/rhysd/github-action-benchmark/actions?query=workflow%3A%22Rust+Example%22)
+-   [Workflow for this example](../../.github/workflows/criterion-rs.yml)
 -   [Benchmark results on GitHub pages](https://rhysd.github.io/github-action-benchmark/dev/bench/)
 
 This directory shows how to use [`github-action-benchmark`](https://github.com/rhysd/github-action-benchmark)
@@ -17,17 +16,11 @@ e.g.
 
 ```yaml
 - name: Run benchmark
-  run: cargo +nightly bench | tee output.txt
+  run: cargo bench -- --output-format bencher | tee output.txt
 ```
 
-Note that this example does not use LTO for benchmarking because entire code in benchmark iteration
-will be removed as dead code. For normal use case, please enable it in `Cargo.toml` for production
-performance.
+Note that you should run the benchmarks using the bencher output format.
 
-```yaml
-[profile.bench]
-lto = true
-```
 
 ## Process benchmark results
