+++ src/default_index_html.ts
@@ -112,7 +112,6 @@ export const DEFAULT_INDEX_HTML = String.raw`<!DOCTYPE html>
         // Colors from https://github.com/github/linguist/blob/master/lib/linguist/languages.yml
         const toolColors = {
           cargo: '#dea584',
-          criterion: '#dea584',
           go: '#00add8',
           benchmarkjs: '#f1e05a',
           pytest: '#3572a5',
